clippy

+++ crates/nu-command/src/system/ps.rs
@@ -132,11 +132,8 @@ fn run_ps(engine_state: &EngineState, call: &Call) -> Result<PipelineData, Shell
                 // record.push("session_id", Value::int(proc_stat.session as i64, span));
                 // This may be helpful for ctrl+z type of checking, once we get there
                 // record.push("tpg_id", Value::int(proc_stat.tpgid as i64, span));
-                record.push("priority", Value::int(proc_stat.priority as i64, span));
-                record.push(
-                    "process_threads",
-                    Value::int(proc_stat.num_threads as i64, span),
-                );
+                record.push("priority", Value::int(proc_stat.priority, span));
+                record.push("process_threads", Value::int(proc_stat.num_threads, span));
             }
             #[cfg(windows)]
             {
